+++ frontend/assets/scss/core/design_system.scss
@@ -56,9 +56,9 @@
 
   &.secondary {
     background-color: black;
-    color:#D59FC7 ;
+    color: #d59fc7;
     position: relative;
-    border:solid 2px $pink ;
+    border: solid 2px $pink;
 
     svg {
       position: absolute;
@@ -70,5 +70,69 @@
       color: $purple;
       transition: all 0.2s ease-in;
     }
+  }
+}
+
+.button-bulle {
+  width: 56px;
+  height: 56px;
+  box-shadow: 0px 3px 6px rgba(0, 0, 0, 0.1), 0px 4px 8px rgba(0, 0, 0, 0.08), 0px 1px 12px rgba(0, 0, 0, 0.04);
+  border: none;
+  border-radius: 24px;
+  cursor: pointer;
+
+  svg {
+    width: 15px;
+    height: 15px;
+  }
+
+  &:hover {
+    box-shadow: 0px 1px 3px rgba(0, 0, 0, 0.15), 0px 2px 4px rgba(0, 0, 0, 0.12);
+  }
+
+  &:active {
+    box-shadow: 0px 0px 2px rgba(0, 0, 0, 0.1);
+  }
+
+  &.purple {
+    background: $purple;
+    color: #ffffff;
+  }
+
+  &.pink {
+    background: $pink;
+    color: #ffffff;
+  }
+}
+
+.button-square {
+  width: 40px;
+  height: 40px;
+  box-shadow: 0px 3px 6px rgba(0, 0, 0, 0.1), 0px 4px 8px rgba(0, 0, 0, 0.08), 0px 1px 12px rgba(0, 0, 0, 0.04);
+  border: none;
+  border-radius: 12px;
+  cursor: pointer;
+
+  svg {
+    width: 15px;
+    height: 15px;
+  }
+
+  &:hover {
+    box-shadow: 0px 1px 3px rgba(0, 0, 0, 0.15), 0px 2px 4px rgba(0, 0, 0, 0.12);
+  }
+
+  &:active {
+    box-shadow: 0px 0px 2px rgba(0, 0, 0, 0.1);
+  }
+
+  &.purple {
+    background: $purple;
+    color: #ffffff;
+  }
+
+  &.pink {
+    background: $pink;
+    color: #ffffff;
   }
 }
