+++ frontend/assets/scss/vinyles_store/Homepage.module.scss
@@ -1,0 +1,116 @@
+@import "../core/variables.scss";
+
+.content {
+  height: 100vh;
+  background-color: $purple;
+}
+
+.header {
+  display: flex;
+  align-items: center;
+  background: #ffffff;
+  box-shadow: 0px 4px 5px rgba(255, 255, 255, 0.25);
+  gap: 20px;
+  padding: 30px 20px;
+
+  span {
+    font-family: "Wedges", sans-serif;
+    font-size: 3.2rem;
+    color: $purple;
+  }
+
+  form {
+    display: flex;
+    align-items: center;
+    margin-left: auto;
+
+    input {
+      width: 400px;
+      border-top-right-radius: 0;
+      border-bottom-right-radius: 0;
+    }
+
+    button {
+      height: 56px;
+      background-color: $purple;
+      color: #ffffff;
+      border: none;
+      border: 1px solid #e5e6e7;
+      border-radius: 8px;
+      border-top-left-radius: 0;
+      border-bottom-left-radius: 0;
+      font-weight: 400;
+      font-size: 1.6rem;
+      padding: 10px 20px;
+      cursor: pointer;
+
+      svg {
+        width: 18px;
+        height: 18px;
+      }
+    }
+  }
+}
+
+.main {
+  display: grid;
+  grid-template-columns: repeat(12, 1fr);
+  grid-gap: 10px;
+  padding: 50px;
+
+  h1 {
+    grid-column: 1 / 13;
+    grid-row: 1;
+    font-family: "Wedges", sans-serif;
+    font-size: 3.2rem;
+    color: #ffffff;
+    margin-bottom: 50px;
+  }
+
+  .wantlist {
+    grid-column: 1 / 13;
+    display: grid;
+    grid-template-columns: repeat(12, 1fr);
+    gap: 20px;
+
+    & > div {
+      grid-column: span 3;
+      background-color: #ffffff;
+      display: flex;
+      flex-direction: column;
+      padding: 20px;
+
+      & > div:first-child {
+        height: 165px;
+        background: #ffffff url("../../img/background-vinyle.png") no-repeat;
+        background-size: contain;
+
+        img {
+          width: 130px;
+          height: 130px;
+          position: relative;
+          top: 10px;
+          left: 10px;
+        }
+      }
+
+      h2 {
+        font-weight: 600;
+        font-size: 2rem;
+        line-height: 24px;
+        color: $purple;
+      }
+
+      p {
+        color: #acacac;
+        margin: 10px 0 20px;
+      }
+
+      & > div:last-child {
+        button {
+          margin-right: 10px;
+        }
+      }
+    }
+  }
+}
